The following project aims to be an Air bnb like app
The android application has 2 side, the side of the home owner and the side of the user that is looking for home to rent.
The user that logs as home owner can upload an ad with pictures and map location and the dates available.
The user that logs as someone how is looking for a home to rent can find rooms based on location rating and price